# storage: optimize `pointSynthesizingIter` for scans